+++ src/api/paperless.ts
@@ -16,6 +16,7 @@ import type {
 	TagRequest,
 	UpdateCorrespondentRequest,
 	UpdateDocumentRequest,
+	UpdateDocumentTypeRequest,
 } from '#types';
 
 export class PaperlessAPI {
@@ -258,6 +259,16 @@ export class PaperlessAPI {
 		return this.request<DocumentType>('/document_types/', {
 			method: 'POST',
 			body: JSON.stringify(data),
+		});
+	}
+
+	async updateDocumentType(
+		id: number,
+		data: UpdateDocumentTypeRequest,
+	): Promise<DocumentType> {
+		return this.request<DocumentType>(`/document_types/${id}/`, {
+			method: 'PATCH',
+			body: JSON.stringify(omitUndefined(data)),
 		});
 	}